testing/web-platform/tests/encoding-detection/ko-EUC-KR.tentative.html

<!doctype html>
<title>ko EUC-KR</title>
<script src=/resources/testharness.js></script>
<script src=/resources/testharnessreport.js></script>
<p>�̰��� ���� ���ڵ� �׽�Ʈ�Դϴ�.</p>
<script>
setup({explicit_done:true});
onload = function() {
    test(function() {
        assert_equals(document.characterSet, "EUC-KR", 'Expected EUC-KR');
    }, "Check detection result");
      done();
};
</script>
